Move getLyricUpdateCommand into lyrics service and rename
